1. LookupCalculationLogic 计算逻辑实现. 2. 修正关于ID生成的生成器,统一使用common/id包中的相关实现.

+++ xplat-meta-oqsengine-boot/src/main/java/com/xforceplus/ultraman/oqsengine/boot/grpc/service/EntityRebuildServiceOqs.java
@@ -344,10 +344,10 @@ public class EntityRebuildServiceOqs implements EntityRebuildServicePowerApi {
             .setIsCancel(taskInfo.isCancel())
             .setIsDone(taskInfo.isDone())
             .setStatus(taskInfo.status().name())
-            .setBatchSize(taskInfo.getBatchSize())
+            .setBatchSize((int) taskInfo.getBatchSize())
             .setEntityId(taskInfo.getEntity())
             .setFinishSize(taskInfo.getFinishSize())
-            .setPercentage(taskInfo.getProgressPercentage())
+            .setPercentage((int) taskInfo.getProgressPercentage())
             .setStarts(taskInfo.getStarts())
             .setEnds(taskInfo.getEnds())
             .build();
